requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-filter==23.1
djangorestframework-simplejwt
factory_boy~=3.3
